refactor(iast): convert include paths from PascalCase to snake_case
