+++ ddtrace/appsec/_iast/src/taint_tracking/source.cpp
@@ -1,5 +1,6 @@
-#include "source.h"
 #include <pybind11/pybind11.h>
+
+#include "source.h"
 
 using namespace std;
 namespace py = pybind11;
